[misc] Rename the app to infinite-livestream
"livestream" described the medium rather than the thing. The app is the
infinite livestream: a channel that keeps generating whether or not anybody is
watching or typing.

Named the way the monorepo names apps. The directory takes an underscore, as
`apps/fastvideo_studio` and `apps/performance_dashboard` do, so the package is
`infinite_livestream`; the distribution and the console script take the hyphen
that is conventional for both, as `dreamverse-server` does.

  apps/livestream/            -> apps/infinite_livestream/
  livestream/                 -> infinite_livestream/
  livestream-server           -> infinite-livestream-server
  configs/livestream.yaml     -> configs/infinite_livestream.yaml

`main.py` had hardcoded its logger as "livestream" while every other module
derives one from `__name__`, so the top-level line was the only one that would
not have followed the rename.

No behaviour change. 60 tests, lint and mypy clean, and the stream restarted
under the new package: ready in 193.6 s, clips at 1.08-1.14x realtime, zero
dropped frames, +0.000s A/V skew.
