+++ apps/infinite_livestream/infinite_livestream/__init__.py
@@ -1,0 +1,5 @@
+"""Infinite Livestream: a chat-driven FastH3 broadcast."""
+
+__all__ = ["__version__"]
+
+__version__ = "0.1.0"
