--- apps/livestream/livestream/__init__.py
@@ -1,5 +1,0 @@
-"""Chat-driven FastH3 infinite livestream."""
-
-__all__ = ["__version__"]
-
-__version__ = "0.1.0"
